\name{RatingScaleReduction-package}
\alias{RatingScaleReduction-package}
\alias{RatingScaleReduction}
\docType{package}
\title{Rating Scale Reduction Procedure}

\description{
This package describes a procedure of reducing items in a rating scale.
It was published in the refence included in this description. The method was proposed by Waldemar W. Koczkodaj
and published by a sizable collboration coordinated by him.
}

\author{
Waldemar W. Koczkodaj, Feng Li,Alicja Wolny-Dominiak \cr
Maintainer: Alicja Wolny-Dominiak
}


\references{
1. W.W. Koczkodaj,  T. Kakiashvili,  A. Szymanska, J. Montero-Marin, R. Araya, J. Garcia-Campayo, K. Rutkowski, D. Strzalka,
 How to reduce the number of rating scale items without
 predictability loss? Scientometrics, 909(2):581-593(open access), 2017 \cr
 \url{https://link.springer.com/article/10.1007/s11192-017-2283-4} \cr


2. T. Kakiashvili, W. W. Koczkodaj, and M. Woodbury-Smith. Improving the medical scale predictability
by the pairwise comparisons method: Evidence from a clinical data study. Computer Methods and
Programs in Biomedicine, 105(3), 2012 \cr
DOI: \url{http://dx.doi.org/10.1016/j.cmpb.2011.09.011} \cr


3. X. Robin, N. Turck, A. Hainard, N. Tiberti, F. Lisacek, J.-C. Sanchez, and M. Muller. proc: an opensource
package for r and s+ to analyze and compare roc curves. BMC Bioinformatics, 2011 \cr
\url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-12-77} \cr

4. R. DeLong, D. M. DeLong, and D. L. Clarke-Pearson. Comparing the areas under two or more
correlated receiver operating characteristic curves: a nonparametric approach. Biometrics, 1988 \cr
}
