\name{SHSData}
\alias{SHSData}
\docType{data}
\title{
shsData}
\description{
.............}
\usage{data("SHSData")}
\format{
  A data frame with 143 observations on the following 7 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{D8}}{a numeric vector}
  }
}
\references{
.......................................}
\examples{
data(SHSData)
## maybe str(SHSData) ; plot(SHSData) ...
}
\keyword{datasets}
