\name{wineData}
\alias{wineData}
\docType{data}
\title{
.................
}
\description{
......................}
\usage{data("wineData")}
\format{
  A data frame with 6497 observations on the following 13 variables.
  \describe{
    \item{\code{fixed.acidity}}{a numeric vector}
    \item{\code{volatile.acidity}}{a numeric vector}
    \item{\code{citric.acid}}{a numeric vector}
    \item{\code{residual.sugar}}{a numeric vector}
    \item{\code{chlorides}}{a numeric vector}
    \item{\code{free.sulfur.dioxide}}{a numeric vector}
    \item{\code{total.sulfur.dioxide}}{a numeric vector}
    \item{\code{density}}{a numeric vector}
    \item{\code{pH}}{a numeric vector}
    \item{\code{sulphates}}{a numeric vector}
    \item{\code{alcohol}}{a numeric vector}
    \item{\code{quality}}{a numeric vector}
  }
}
\details{
.............................}

\references{
........................................
}
\examples{
data(wineData)
## maybe str(wineData) ; plot(wineData) ...
}
\keyword{datasets}
