\name{RatingScaleReduction-package}
\alias{RatingScaleReduction-package}
\alias{RatingScaleReduction}
\docType{package}
\title{Rating Scale Reduction Procedure}

\description{
In essence, this package describes a procedure of reducing items in a rating scale.
It was published in the refence included in this description. The method was proposed by Waldemar W. Koczkodaj
and published by a sizable collboration coordinated by him.
}

\references{W.W. Koczkodaj,  T. Kakiashvili,  A. Szymanska, 
 J. Montero-Marin, R. Araya, J. Garcia-Campayo, K. Rutkowski,
 D. Strzalka (2017)
 How to reduce the number of rating scale items without
 predictability loss? Scientometrics. <https://arxiv.org/pdf/1505.01886.pdf>}
